+++ modules/module-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @powersync/service-module-core
 
+## 0.2.30
+
+### Patch Changes
+
+- Updated dependencies [71d4a0a]
+- Updated dependencies [e4f683d]
+- Updated dependencies [71d4a0a]
+- Updated dependencies [c2edf86]
+- Updated dependencies [df9ab1e]
+  - @powersync/service-core@1.23.2
+  - @powersync/service-types@0.16.1
+  - @powersync/lib-services-framework@0.9.7
+  - @powersync/service-rsocket-router@0.2.23
+
 ## 0.2.29
 
 ### Patch Changes